+++ lib/boot.dart
@@ -1,2 +1,2 @@
-export 'src/boot/boot_state.dart';
 export 'src/boot/boot_scope.dart';
+export 'src/boot/boot_state.dart';