+++ lib/chain.dart
@@ -1,2 +1,3 @@
-export 'src/widget/chain/widgets.dart';
 export 'src/widget/chain/painting.dart';
+export 'src/widget/chain/widgets.dart';
+